+++ daisylib-api/src/main/java/uk/firedev/daisylib/api/database/DatabaseModule.java
@@ -1,4 +1,4 @@
-package uk.firedev.daisylib.database;
+package uk.firedev.daisylib.api.database;
 
 import org.jetbrains.annotations.NotNull;
